requirements.txt
PyYAML==6.0.1
openai==0.27.8
torch==2.0.1
numpy==1.24.4
wandb==0.15.8
tqdm==4.65.0
opencv-python==4.8.0.74
pandas==2.0.3
torchvision==0.15.2
bosdyn-api==3.3.0
bosdyn-choreography-client==3.3.0
bosdyn-choreography-protos==3.3.0
bosdyn-client==3.3.0
bosdyn-core==3.3.0
bosdyn-mission==3.3.0
black==24.2.0
isort==5.12.0
pylint==2.17.5
open3d==0.16.0
Pillow==10.0.0
scipy==1.10.1
matplotlib==3.7.2
apriltag==0.0.16
ikpy==3.3.4
Deprecated==1.2.14
pdf2image==1.16.3
Flask==2.2.5
requests==2.31.0
transformers==4.33.3
urllib3==2.0.4